Add pre-commit tooling (#12)
This PR adds and rolls out `lefthook` for our `pre-commit` tool. It's
something I'm keen to try out, so here we go.

<!-- av pr metadata
This information is embedded by the av CLI when creating PRs to track
the status of stacks when using Aviator. Please do not delete or edit
this section of the PR.
```
{"parent":"main","parentHead":"","trunk":"main"}
```
-->